The pages are bytes, and a write that changes nothing dirties nothing
A page held its bytes as ints, four bytes each: a 128K's RAM was half a megabyte, the
screen twenty-seven kilobytes, and all of it competing for the cache with the tables a
frame walks. They are bytes now - RAMHolder, MemoryPage, Memory.newRam, the floating bus,
the snapshot load, the three places Display reads the screen, and the three devices that
fill their own pages - and the core reads them masked and stores them as they are.

Bytes showed up a comparison that had been wrong all along: whether a write changes the
byte was asked as an int from 0 to 255 against a signed byte, so any byte of 0x80 or more
written unchanged counted as changed and was marked dirty - half the pixel bytes of a
sprite, one dirty8 each, or eight marks for an attribute with FLASH. MemoryPage.holds
compares byte with byte.

Measured alternating on the same cores: Manic Miner within noise; Jet Set Willy, which
rewrites its screen every frame, 7882 to 11697 frames a second, forty-eight percent. In
its profile dirty64, which was a third of the samples, is no longer in the top eight.

machine/core 272 on both cores, all of machine built, machine/app 30 with the slow ones,
the generated core current.

+++ machine/core/src/main/java/com/fpetrola/oozx/DefaultRAMHolder.java
@@ -20,9 +20,9 @@ package com.fpetrola.oozx;
 
 public class DefaultRAMHolder implements RAMHolder {
   // RAM array: 65 pages of 16KB each (from SPECTRUM_RAM_PAGES)
-  private final int[][] RAM = new int[Memory.SPECTRUM_RAM_PAGES][0x4000];
+  private final byte[][] RAM = new byte[Memory.SPECTRUM_RAM_PAGES][0x4000];
 
-  public int[][] getRAM() {
+  public byte[][] getRAM() {
     return RAM;
   }
 }
